allow not to build doc when doxygen is not available (useful for binary-only builds)

+++ build/build.sh
@@ -2,8 +2,6 @@
 
 # Called by build tools (Travis...), don't use directly
 
-echo "Building: $1"
-
 cd $(dirname $(readlink -f $0))/..
 
 cp -af build/debian.skel debian
@@ -11,9 +9,12 @@ make target/version
 version=$(< target/version)
 echo version=$version
 
-for f in build/debian.$1/*; do
-    sed "s/#VERSION#/$version/g" $f > debian/${f##*/}
-done
+if [[ -n "$1" ]]; then
+    echo "Building: $1"
+    for f in build/debian.$1/*; do
+        sed "s/#VERSION#/$version/g" $f > debian/${f##*/}
+    done
 
-rm -f debian/files
-exec make release.$1
+    rm -f debian/files
+    exec make release.$1
+fi
